# Conflicts:
#	scripts/fbt_tools/crosscc.py
#	site_scons/commandline.scons

+++ scripts/toolchain/fbtenv.sh
@@ -4,9 +4,15 @@
 
 # public variables
 DEFAULT_SCRIPT_PATH="$(pwd -P)";
-SCRIPT_PATH="${SCRIPT_PATH:-$DEFAULT_SCRIPT_PATH}";
-FBT_TOOLCHAIN_VERSION="${FBT_TOOLCHAIN_VERSION:-"21"}";
-FBT_TOOLCHAIN_PATH="${FBT_TOOLCHAIN_PATH:-$SCRIPT_PATH}";
+FBT_TOOLCHAIN_VERSION="${FBT_TOOLCHAIN_VERSION:-"23"}";
+
+if [ -z ${FBT_TOOLCHAIN_PATH+x} ] ; then
+    FBT_TOOLCHAIN_PATH_WAS_SET=0;
+else
+    FBT_TOOLCHAIN_PATH_WAS_SET=1;
+fi
+
+FBT_TOOLCHAIN_PATH="${FBT_TOOLCHAIN_PATH:-$DEFAULT_SCRIPT_PATH}";
 FBT_VERBOSE="${FBT_VERBOSE:-""}";
 
 fbtenv_show_usage()
@@ -32,11 +38,11 @@ fbtenv_wget()
 fbtenv_restore_env()
 {
     TOOLCHAIN_ARCH_DIR_SED="$(echo "$TOOLCHAIN_ARCH_DIR" | sed 's/\//\\\//g')"
-    PATH="$(echo "$PATH" | /usr/bin/sed "s/$TOOLCHAIN_ARCH_DIR_SED\/python\/bin://g")";
-    PATH="$(echo "$PATH" | /usr/bin/sed "s/$TOOLCHAIN_ARCH_DIR_SED\/bin://g")";
-    PATH="$(echo "$PATH" | /usr/bin/sed "s/$TOOLCHAIN_ARCH_DIR_SED\/protobuf\/bin://g")";
-    PATH="$(echo "$PATH" | /usr/bin/sed "s/$TOOLCHAIN_ARCH_DIR_SED\/openocd\/bin://g")";
-    PATH="$(echo "$PATH" | /usr/bin/sed "s/$TOOLCHAIN_ARCH_DIR_SED\/openssl\/bin://g")";
+    PATH="$(echo "$PATH" | sed "s/$TOOLCHAIN_ARCH_DIR_SED\/python\/bin://g")";
+    PATH="$(echo "$PATH" | sed "s/$TOOLCHAIN_ARCH_DIR_SED\/bin://g")";
+    PATH="$(echo "$PATH" | sed "s/$TOOLCHAIN_ARCH_DIR_SED\/protobuf\/bin://g")";
+    PATH="$(echo "$PATH" | sed "s/$TOOLCHAIN_ARCH_DIR_SED\/openocd\/bin://g")";
+    PATH="$(echo "$PATH" | sed "s/$TOOLCHAIN_ARCH_DIR_SED\/openssl\/bin://g")";
     if [ -n "${PS1:-""}" ]; then
         PS1="$(echo "$PS1" | sed 's/\[fbt\]//g')";
     elif [ -n "${PROMPT:-""}" ]; then
@@ -50,6 +56,16 @@ fbtenv_restore_env()
         unset SSL_CERT_FILE;
         unset REQUESTS_CA_BUNDLE;
     fi
+
+    if [ "$SYS_TYPE" = "Linux" ]; then
+        if [ -n "$SAVED_TERMINFO_DIRS" ]; then
+            export TERMINFO_DIRS="$SAVED_TERMINFO_DIRS";
+        else
+            unset TERMINFO_DIRS;
+        fi
+        unset SAVED_TERMINFO_DIRS;
+    fi
+
     export PYTHONNOUSERSITE="$SAVED_PYTHONNOUSERSITE";
     export PYTHONPATH="$SAVED_PYTHONPATH";
     export PYTHONHOME="$SAVED_PYTHONHOME";
@@ -60,13 +76,15 @@ fbtenv_restore_env()
     unset SAVED_PYTHONPATH;
     unset SAVED_PYTHONHOME;
 
-    unset SCRIPT_PATH;
     unset FBT_TOOLCHAIN_VERSION;
     unset FBT_TOOLCHAIN_PATH;
 }
 
 fbtenv_check_sourced()
 {
+    if [ -n "${FBT_SKIP_CHECK_SOURCED:-""}" ]; then
+        return 0;
+    fi
     case "${ZSH_EVAL_CONTEXT:-""}" in *:file:*)
         setopt +o nomatch;  # disabling 'no match found' warning in zsh
         return 0;;
@@ -89,8 +107,6 @@ fbtenv_check_if_sourced_multiple_times()
             return 0;
         fi
     fi
-    echo "Warning! FBT environment script was sourced more than once!";
-    echo "You might be doing things wrong, please open a new shell!";
     return 1;
 }
 
@@ -104,13 +120,14 @@ fbtenv_set_shell_prompt()
     return 0;  # all other shells
 }
 
-fbtenv_check_script_path()
+fbtenv_check_env_vars()
 {
-    if [ ! -x "$SCRIPT_PATH/fbt" ] && [ ! -x "$SCRIPT_PATH/ufbt" ] ; then
-        echo "Please source this script from [u]fbt root directory, or specify 'SCRIPT_PATH' variable manually";
+    # Return error if FBT_TOOLCHAIN_PATH is not set before script is sourced or if fbt executable is not in DEFAULT_SCRIPT_PATH
+    if [ "$FBT_TOOLCHAIN_PATH_WAS_SET" -eq 0 ] && [ ! -x "$DEFAULT_SCRIPT_PATH/fbt" ] && [ ! -x "$DEFAULT_SCRIPT_PATH/ufbt" ] ; then
+        echo "Please source this script from [u]fbt root directory, or specify 'FBT_TOOLCHAIN_PATH' variable manually";
         echo "Example:";
-        printf "\tSCRIPT_PATH=lang/c/flipperzero-firmware source lang/c/flipperzero-firmware/scripts/fbtenv.sh\n";
-        echo "If current directory is right, type 'unset SCRIPT_PATH' and try again"
+        printf "\tFBT_TOOLCHAIN_PATH=lang/c/flipperzero-firmware source lang/c/flipperzero-firmware/scripts/fbtenv.sh\n";
+        echo "If current directory is right, type 'unset FBT_TOOLCHAIN_PATH' and try again"
         return 1;
     fi
     return 0;
@@ -144,7 +161,7 @@ fbtenv_get_kernel_type()
 fbtenv_check_rosetta()
 {
     if [ "$ARCH_TYPE" = "arm64" ]; then
-        if ! /usr/bin/pgrep -q oahd; then
+        if ! pgrep -q oahd; then
             echo "Flipper Zero Toolchain needs Rosetta2 to run under Apple Silicon";
             echo "Please instal it by typing 'softwareupdate --install-rosetta --agree-to-license'";
             return 1;
@@ -186,7 +203,7 @@ fbtenv_download_toolchain_tar()
     return 0;
 }
 
-fbtenv_remove_old_tooclhain()
+fbtenv_remove_old_toolchain()
 {
     printf "Removing old toolchain..";
     rm -rf "${TOOLCHAIN_ARCH_DIR:?}";
@@ -207,7 +224,7 @@ fbtenv_show_unpack_percentage()
 
 fbtenv_unpack_toolchain()
 {
-    echo "Unpacking toolchain:";
+    echo "Unpacking toolchain to '$FBT_TOOLCHAIN_PATH/toolchain':";
     tar -xvf "$FBT_TOOLCHAIN_PATH/toolchain/$TOOLCHAIN_TAR" -C "$FBT_TOOLCHAIN_PATH/toolchain" 2>&1 | fbtenv_show_unpack_percentage;
     mkdir -p "$FBT_TOOLCHAIN_PATH/toolchain" || return 1;
     mv "$FBT_TOOLCHAIN_PATH/toolchain/$TOOLCHAIN_DIR" "$TOOLCHAIN_ARCH_DIR" || return 1;
@@ -215,14 +232,16 @@ fbtenv_unpack_toolchain()
     return 0;
 }
 
-fbtenv_clearing()
+fbtenv_cleanup()
 {
-    printf "Cleaning up..";
     if [ -n "${FBT_TOOLCHAIN_PATH:-""}" ]; then
-        rm -rf "${FBT_TOOLCHAIN_PATH:?}/toolchain/"*.tar.gz;
+        printf "Cleaning up..";
         rm -rf "${FBT_TOOLCHAIN_PATH:?}/toolchain/"*.part;
+        if [ -z "${FBT_PRESERVE_TAR:-""}" ]; then
+            rm -rf "${FBT_TOOLCHAIN_PATH:?}/toolchain/"*.tar.gz;
+        fi
+        echo "done";
     fi
-    echo "done";
     trap - 2;
     return 0;
 }
@@ -270,21 +289,27 @@ fbtenv_download_toolchain()
     fbtenv_check_tar || return 1;
     TOOLCHAIN_TAR="$(basename "$TOOLCHAIN_URL")";
     TOOLCHAIN_DIR="$(echo "$TOOLCHAIN_TAR" | sed "s/-$FBT_TOOLCHAIN_VERSION.tar.gz//g")";
-    trap fbtenv_clearing 2;  # trap will be restored in fbtenv_clearing
+    trap fbtenv_cleanup 2;  # trap will be restored in fbtenv_cleanup
     if ! fbtenv_check_downloaded_toolchain; then
         fbtenv_curl_wget_check || return 1;
         fbtenv_download_toolchain_tar || return 1;
     fi
-    fbtenv_remove_old_tooclhain;
+    fbtenv_remove_old_toolchain;
     fbtenv_unpack_toolchain || return 1;
-    fbtenv_clearing;
+    fbtenv_cleanup;
     return 0;
 }
 
-fbtenv_print_version()
+fbtenv_print_config()
 {
-    if [ -n "$FBT_VERBOSE" ]; then
+    if [ -n "${FBT_VERBOSE:-""}" ]; then
         echo "FBT: using toolchain version $(cat "$TOOLCHAIN_ARCH_DIR/VERSION")";
+        if [ -n "${FBT_SKIP_CHECK_SOURCED:-""}" ]; then
+            echo "FBT: fbtenv will not check if it is sourced or not";
+        fi
+        if [ -n "${FBT_PRESERVE_TAR:-""}" ]; then
+            echo "FBT: toolchain archives will be saved";
+        fi
     fi
 }
 
@@ -296,11 +321,13 @@ fbtenv_main()
         fbtenv_restore_env;
         return 0;
     fi
-    fbtenv_check_if_sourced_multiple_times;  # many source it's just a warning
-    fbtenv_check_script_path || return 1;
+    if ! fbtenv_check_if_sourced_multiple_times; then
+        return 0;
+    fi;
+    fbtenv_check_env_vars || return 1;
     fbtenv_check_download_toolchain || return 1;
     fbtenv_set_shell_prompt;
-    fbtenv_print_version;
+    fbtenv_print_config;
     PATH="$TOOLCHAIN_ARCH_DIR/python/bin:$PATH";
     PATH="$TOOLCHAIN_ARCH_DIR/bin:$PATH";
     PATH="$TOOLCHAIN_ARCH_DIR/protobuf/bin:$PATH";
@@ -319,6 +346,11 @@ fbtenv_main()
     export PYTHONNOUSERSITE=1;
     export PYTHONPATH=;
     export PYTHONHOME=;
+
+    if [ "$SYS_TYPE" = "Linux" ]; then
+        export SAVED_TERMINFO_DIRS="${TERMINFO_DIRS:-""}";
+        export TERMINFO_DIRS="$TOOLCHAIN_ARCH_DIR/ncurses/share/terminfo";
+    fi
 }
 
 fbtenv_main "${1:-""}";
